Split runner.go into specific files and do go fmt on everything
