+++ runner/body.go
@@ -1,0 +1,55 @@
+package runner
+
+import (
+	"fmt"
+	"io"
+	"os"
+	"strings"
+)
+
+type Body struct {
+	File   *os.File
+	Reader io.Reader
+	Offset int64
+	Size   int64
+}
+
+// Create a body
+func NewBody(filename string, offset, size int64) (*Body, error) {
+	file, err := os.Open(filename)
+
+	if err != nil {
+		panic(err)
+	}
+
+	newBody := Body{file, nil, offset, size}
+
+	// Rather than implementing this again
+	newBody.Reset()
+
+	return &newBody, nil
+}
+
+func (b *Body) Reset() error {
+	if _, err := b.File.Seek(b.Offset, 0); err != nil {
+		panic(err)
+	}
+
+	b.Reader = io.LimitReader(b.File, b.Size)
+	return nil
+}
+
+func (b Body) Read(p []byte) (int, error) {
+	return b.Reader.Read(p)
+}
+
+func (b *Body) Close() error {
+	if err := b.File.Close(); err != nil {
+		panic(err)
+	}
+
+	b.File = nil
+	b.Reader = nil
+
+	return nil
+}
